fix: generation scipt complete
