+++ packages/evm-helpers/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/contract';
 export * from './lib/types';
 export * from './lib/metadata';
+export * from './lib/events';
